refactor: 사칙연산 validtor 제거
