+++ src/test/kotlin/step2/inputValidator/ValidatorsTest.kt
@@ -2,13 +2,49 @@ package step2.inputValidator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import step2.inputValidator.FourBasicOperationsValidators.validStartsAndEndsWithNumber
-import step2.inputValidator.FourBasicOperationsValidators.validateNumberInExpression
-import step2.inputValidator.FourBasicOperationsValidators.validateOperatorInExpression
+import step2.inputValidator.InputValidateException.InputIsEmptyException
+import step2.inputValidator.InputValidateException.InputIsNullException
 import step2.inputValidator.InputValidateException.InvalidOperatorException
 import step2.inputValidator.InputValidateException.NonNumericAtExpectedPositionException
+import step2.inputValidator.Validators.validEmptyString
+import step2.inputValidator.Validators.validNull
+import step2.inputValidator.Validators.validStartsAndEndsWithNumber
+import step2.inputValidator.Validators.validateNumberInExpression
+import step2.inputValidator.Validators.validateOperatorInExpression
 
-class FourBasicOperationsValidatorsTest {
+class ValidatorsTest {
+    @Test
+    fun `계산식이 null로 들어올시 Result 실패 상황(InputIsNullException) 을 반환한다`() {
+        val input: List<String>? = null
+
+        assertThat(input.validNull().isFailure).isTrue()
+        assertThat(input.validNull().exceptionOrNull() is InputIsNullException).isTrue()
+    }
+
+    @Test
+    fun `계산식이 null이 아닐시 Result 성공 상황(입력된 값) 을 반환한다`() {
+        val input: List<String> = "1 + 1".split(" ")
+
+        assertThat(input.validNull().isSuccess).isTrue()
+        assertThat(input.validNull().getOrNull() == "1 + 1".split(" ")).isTrue()
+    }
+
+    @Test
+    fun `계산식이 빈스트링으로 들어올시 Result 실패 상황(InputIsEmptyException) 을 반환한다`() {
+        val input: List<String> = "     ".split(" ")
+
+        assertThat(input.validEmptyString().isFailure).isTrue()
+        assertThat(input.validEmptyString().exceptionOrNull() is InputIsEmptyException).isTrue()
+    }
+
+    @Test
+    fun `계산식이 빈스트링이 아닐시 Result 성공 상황(입력된 값) 을 반환한다`() {
+        val input: List<String> = "1 + 1".split(" ")
+
+        assertThat(input.validEmptyString().isSuccess).isTrue()
+        assertThat(input.validEmptyString().getOrNull() == "1 + 1".split(" ")).isTrue()
+    }
+
     @Test
     fun `계산식이 숫자로 시작하지 않을 경우 Result 실패 상황(NonNumericStartOrEndException) 을 반환한다`() {
         val input: List<String> = "d + 1".split(" ")
